Cleaned up the initial data read to skip unused columns and rows.
Changed it so that the column is called 'Product' at each stage of the process.

Cleaned up inflation plot function to be simpler with layout done after the correct y axis variable has been selected.

Made a new function which combines the data cleaning and data plotting into one so that it can be one line in the app.